lib vtls schannel_verify

+++ lib/vtls/schannel_verify.c
@@ -118,7 +118,7 @@ static CURLcode add_certs_data_to_store(HCERTSTORE trust_store,
 
   while(more_certs && (current_ca_file_ptr<ca_buffer_limit)) {
     const char *begin_cert_ptr = c_memmem(current_ca_file_ptr,
-                                          ca_buffer_limit-current_ca_file_ptr,
+                               (size_t)(ca_buffer_limit - current_ca_file_ptr),
                                           BEGIN_CERT,
                                           begin_cert_len);
     if(!begin_cert_ptr || !is_cr_or_lf(begin_cert_ptr[begin_cert_len])) {
@@ -126,7 +126,7 @@ static CURLcode add_certs_data_to_store(HCERTSTORE trust_store,
     }
     else {
       const char *end_cert_ptr = c_memmem(begin_cert_ptr,
-                                          ca_buffer_limit-begin_cert_ptr,
+                                    (size_t)(ca_buffer_limit - begin_cert_ptr),
                                           END_CERT,
                                           end_cert_len);
       if(!end_cert_ptr) {
